requirements.txt
Django==5.1
python-decouple==3.8
psycopg==3.2.1
psycopg-binary==3.2.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
setuptools==72.2.0
django-redis==5.4.0
coverage==7.6.1
django-debug-toolbar==4.4.6